draftclaw/requirements.txt
requests>=2.28.0
python-dotenv>=1.0.0
urllib3>=2.0.0
Pillow>=10.0.0
PyMuPDF>=1.23.0
